+++ content_settings/conf.py
@@ -13,7 +13,7 @@ It has the following global variables:
 
 from importlib import import_module
 from functools import partial
-from typing import Any, Callable, Optional, List, Set, Dict, Tuple
+from typing import Any, Callable, Optional, List, Set, Dict, Tuple, Union
 from functools import cached_property
 
 from django.apps import apps
@@ -355,6 +355,22 @@ class _Settings:
             ), f"Invalid attribute name: {value}; prefix not found"
             return PREFIXSES[prefix](name, suffix)
         return get_value(name, suffix)
+
+    def __setattr__(self, name: str, value: Union[str, Tuple]) -> None:
+        prefix, name, suffix = split_attr(name)
+        assert not prefix, "Can not set attribute with prefix"
+        assert not suffix, "Can not set attribute with suffix"
+        assert isinstance(value, str) or (
+            isinstance(value, tuple) and USER_DEFINED_TYPES
+        ), "You can set only raw value or user defined type"
+
+        from .caching import set_new_db_value
+
+        if isinstance(value, str):
+            set_new_db_value(name, value)
+        else:
+            assert USER_DEFINED_TYPES, "User defined types are not enabled"
+            set_new_db_value(name, *value)
 
     def __dir__(self):
         """